feat: Add success state to Mini Tax Station Registration page
The code changes in this commit add a new `success` state to the Mini Tax Station Registration page. This state is used to track whether the registration process was successful or not. It is set to `false` by default and can be updated based on the outcome of the registration request.

This addition allows for better user feedback and helps to differentiate between successful and failed registration attempts.